feat: rearrange dpm test code
