+++ tests/simulate/report/archmagefb/test_overview.py
@@ -2,54 +2,8 @@ import simaple.simulate.component.skill  # pylint: disable=W0611
 from simaple.core.base import Stat
 from simaple.simulate.actor import time_elapsing_action
 from simaple.simulate.base import Action
-from simaple.simulate.component.view import BuffParentView, Validity, ValidityParentView
+from simaple.simulate.component.view import BuffParentView
 from simaple.simulate.reserved_names import Tag
-
-
-def test_archmage_fb(archmagefb_client):
-
-    schedule = [
-        ("메디테이션", "use", None),
-        ("이프리트", "use", None),
-        ("퓨리 오브 이프리트", "use", None),
-        ("포이즌 체인", "use", None),
-        ("플레임 헤이즈", "use", None),
-        ("미스트 이럽션", "use", None),
-        ("플레임 헤이즈", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("미스트 이럽션", "use", None),
-        ("플레임 헤이즈", "use", None),
-        ("도트 퍼니셔", "use", None),
-        ("포이즌 노바", "use", None),
-        ("플레임 스윕", "use", None),
-        ("플레임 스윕", "use", None),
-        ("미스트 이럽션", "use", None),
-        ("플레임 헤이즈", "use", None),
-    ]
-
-    actions = [
-        Action(name=name, method=method, payload=payload)
-        for name, method, payload in schedule
-    ]
-
-    for action in actions:
-        events = archmagefb_client.play(action)
-
-
-def test_validity_view(archmagefb_client):
-    archmagefb_client.environment.add_view(
-        "validity", ValidityParentView.build(archmagefb_client.environment)
-    )
-
-    views = archmagefb_client.environment.show("validity")
-
-    for v in views:
-        assert isinstance(v, Validity)
 
 
 def test_buff_view(archmagefb_client):
