+++ tests/simulate/report/bishop/test_bishop.py
